+++ examples/work.cpp
@@ -21,8 +21,8 @@ zth_fiber(job);
 static void handler(int /*sig*/)
 {
 	char const* msg = "Got interrupted. Ignored. Press Ctrl+D to stop.\n";
-	for(ssize_t len = strlen(msg), c = 1; c > 0 && len > 0;
-	    c = write(fileno(stderr), msg, len), len -= c, msg += c)
+	for(ssize_t len = (ssize_t)strlen(msg), c = 1; c > 0 && len > 0;
+	    c = write(fileno(stderr), msg, (size_t)len), len -= c, msg += c)
 		;
 }
 #endif
@@ -69,7 +69,7 @@ static void employer()
 			}
 		}
 
-		len += cnt;
+		len += (size_t)cnt;
 		zth_assert(len < sizeof(buf));
 
 		// Make sure the string is terminated.
@@ -87,10 +87,10 @@ static void employer()
 
 			switch(sscanf(s, "%d%n", &work, &scanned)) {
 			case 1:
-				if(totalScanned + scanned < len) {
+				if(totalScanned + (size_t)scanned < len) {
 					// More chars are in the buffer (at least a whitespace).
 					s += scanned;
-					totalScanned += scanned;
+					totalScanned += (size_t)scanned;
 
 					// Start job.
 					async job(work);
