+++ examples/zmq.cpp
@@ -66,8 +66,8 @@ zth_fiber(client);
 static void handler(int UNUSED_PAR(sig))
 {
 	char const* msg = "got interrupted\n";
-	for(ssize_t len = strlen(msg), c = 1; c > 0 && len > 0;
-	    c = write(fileno(stderr), msg, len), len -= c, msg += c)
+	for(ssize_t len = (ssize_t)strlen(msg), c = 1; c > 0 && len > 0;
+	    c = write(fileno(stderr), msg, (size_t)len), len -= c, msg += c)
 		;
 	stop = 1;
 }
